requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# AI Provider Support
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0

# Data processing and analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For parquet file support
pyarrow>=12.0.0
